+++ packages/gamut/src/Icon/icons/PhoneIcon.tsx
@@ -2,7 +2,7 @@ import React, { SVGProps } from 'react';
 
 import { defaultIconProps } from '../defaultIconProps';
 
-export const PhoneIcon: React.FC<SVGProps<SVGSVGElement>> = props => {
+export const PhoneIcon: React.FC<SVGProps<SVGSVGElement>> = (props) => {
   return (
     <svg {...props}>
       <title>Phone Icon</title>