+++ packages/gamut/src/Tabs/Tabs/index.tsx
@@ -45,9 +45,7 @@ export class Tabs extends Component<TabsProps> {
     }
   }
 
-  idPrefix = Math.random()
-    .toString()
-    .replace('.', '');
+  idPrefix = Math.random().toString().replace('.', '');
 
   createBaseId = (index: number) => `${this.idPrefix}-${index}`;
 
@@ -75,13 +73,13 @@ export class Tabs extends Component<TabsProps> {
     const childrenArray = React.Children.toArray(
       this.props.children
     ) as ReactElement[];
-    let clonedTabPanels = childrenArray.filter(c => c.type === TabPanel);
+    let clonedTabPanels = childrenArray.filter((c) => c.type === TabPanel);
 
     if (activeTabIndex >= clonedTabPanels.length) {
       activeTabIndex = clonedTabPanels.length - 1;
     }
 
-    const tabListChild = childrenArray.find(c => c.type === TabList);
+    const tabListChild = childrenArray.find((c) => c.type === TabList);
     const clonedTabList = React.cloneElement(tabListChild, {
       activeTabIndex,
       onChange,