+++ packages/styleguide/.storybook/webpack.config.js
@@ -27,7 +27,7 @@ module.exports = ({ config }) => {
       includeDefaults: false,
     })
     .css()
-    .if(DEV, config => {
+    .if(DEV, (config) => {
       return config.merge({
         plugins: [new ForkTsCheckerWebpackPlugin()],
       });
@@ -44,10 +44,10 @@ module.exports = ({ config }) => {
   };
 
   const jsIndex = config.module.rules.findIndex(
-    r => r && r.test.test('test.js')
+    (r) => r && r.test.test('test.js')
   );
   if (jsIndex > -1) {
-    config.module.rules[jsIndex].test = /\.(mjs|(j|t)sx?)$/;
+    config.module.rules[jsIndex].test = /\.(mdx|mjs|(j|t)sx?)$/;
     config.module.rules[jsIndex].include.push(
       path.resolve(__dirname, '../../../packages')
     );
@@ -60,6 +60,7 @@ module.exports = ({ config }) => {
       loader: 'react-docgen-typescript-loader',
       options: {
         tsconfigPath: path.resolve(__dirname, '../tsconfig.json'),
+        shouldExtractLiteralValuesFromEnum: true,
         propFilter(p) {
           if (p.parent && p.parent.fileName.match('node_modules')) {
             if (p.required) {
@@ -74,7 +75,7 @@ module.exports = ({ config }) => {
   }
 
   const cssIndex = config.module.rules.findIndex(
-    r => r && r.test.test('test.css')
+    (r) => r && r.test.test('test.css')
   );
   if (cssIndex > -1) {
     config.module.rules[cssIndex] = null;