+++ packages/styleguide/stories/Atoms/IkonaIcon.stories.tsx
@@ -17,8 +17,8 @@ export default {
 };
 
 const iconNames = Object.keys(iconStyles)
-  .filter(cn => cn.match('ikona-'))
-  .map(cn => cn.replace('ikona-', ''));
+  .filter((cn) => cn.match('ikona-'))
+  .map((cn) => cn.replace('ikona-', ''));
 
 export const ikonaIcons = decoratedStory(() => (
   <StoryTemplate status={StoryStatus.Deprecated} wide>
@@ -30,7 +30,7 @@ export const ikonaIcons = decoratedStory(() => (
       <code>@codecademy/gamut-icons</code> package instead.
     </StoryDescription>
     <LayoutGrid columnGap="sm" rowGap="sm">
-      {iconNames.map(iconName => (
+      {iconNames.map((iconName) => (
         <Column size={1} key={iconName}>
           <IkonaIcon key={iconName} name={iconName} size={2} />
           <span>{iconName}</span>