+++ packages/styleguide/stories/Brand/Atoms/Avatar/Avatar.stories.tsx
@@ -18,7 +18,7 @@ export default {
 
 export const avatar = decoratedStory(() => (
   <StoryTemplate status={StoryStatus.Ready}>
-    {theme => (
+    {(theme) => (
       <>
         <StoryDescription>
           Small, circular profile photo that includes a colored drop shadow.