+++ packages/styleguide/stories/Brand/Atoms/Quote.stories.tsx
@@ -17,7 +17,7 @@ export default {
 
 export const quote = decoratedStory(() => (
   <StoryTemplate status={StoryStatus.Ready}>
-    {theme => (
+    {(theme) => (
       <>
         <StoryDescription>
           Emphasized text intended to be a direct quote, such as from a learner.