+++ packages/webpack-config/src/config/css-extracted.js
@@ -2,7 +2,7 @@ const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 const merge = require('webpack-merge');
 const loaders = require('../loaders');
 
-const cssExtracted = options => {
+const cssExtracted = (options) => {
   const defaultExtractOptions = {
     filename: '[name].css',
     esModule: true,